refactor: remove redundant abstraction in the proxy layer
